prevent pdf rendering logs in case of failure

+++ pkg/render/render.go
@@ -242,15 +242,19 @@ func NewPage(cwd string, saveToDisk bool) (*PageOpts, error) {
 	}, nil
 }
 
-func (n *PageOpts) GenerateSinglePageActivity() error {
+func (n *PageOpts) GenerateSinglePageActivity(preventCleanup bool) error {
 	tempDir, err := os.MkdirTemp(".", "tmp")
 	if err != nil {
 		log.Fatal().Err(err).Msg("Failed to create temp directory")
 	}
 
 	defer func() {
-		time.Sleep(1 * time.Second) // Sleep for 1 second before removing the directory
-		os.RemoveAll(tempDir)
+		if !preventCleanup {
+			time.Sleep(1 * time.Second) // Sleep for 1 second before removing the directory
+			os.RemoveAll(tempDir)
+		} else {
+			log.Info().Msgf("Asset rendering folder not removed (%s)", tempDir)
+		}
 	}()
 
 	n.TmpDir = tempDir
